NeoCortex/requirements.txt
numpy
opencv
pyserial
fcntl
msvcrt
tty
argparse
matplotlib




